fix: regenerate SDK with updated OpenAPI spec
Regenerated from spec changes: extracted DirectoryUser/EventContext
schemas, SlimRole refs, limit type number->integer, RoleList ref.

+++ lib/Resource/AuditLogEvent.php
@@ -30,7 +30,7 @@ readonly class AuditLogEvent implements \JsonSerializable
          */
         public ?array $metadata = null,
         /** What schema version the event is associated with. */
-        public ?float $version = null,
+        public ?int $version = null,
     ) {
     }
