+++ lib/Service/ApiKeys.php
@@ -65,7 +65,7 @@ class ApiKeys
      * @param string $organizationId Unique identifier of the Organization.
      * @param string|null $before An object ID that defines your place in the list. When the ID is not present, you are at the end of the list.
      * @param string|null $after An object ID that defines your place in the list. When the ID is not present, you are at the end of the list.
-     * @param float|null $limit Upper limit on the number of objects to return, between `1` and `100`. Defaults to 10.
+     * @param int|null $limit Upper limit on the number of objects to return, between `1` and `100`. Defaults to 10.
      * @param \WorkOS\Resource\EventsOrder|null $order Order the results by the creation time. Defaults to "desc".
      * @return \WorkOS\PaginatedResponse<\WorkOS\Resource\ApiKey>
      */
@@ -73,7 +73,7 @@ class ApiKeys
         string $organizationId,
         ?string $before = null,
         ?string $after = null,
-        ?float $limit = null,
+        ?int $limit = null,
         ?\WorkOS\Resource\EventsOrder $order = null,
         ?\WorkOS\RequestOptions $options = null,
     ): \WorkOS\PaginatedResponse {
